[fix]adde validation open in 005

+++ 005_exercise_heredoc_open/main.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   main.c                                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: chguerre <chguerre@student.lausanne.ch>    +#+  +:+       +#+        */
+/*   By: chguerre <chguerre@student.42lausanne.c    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/09/01 20:00:54 by chguerr           #+#    #+#             */
-/*   Updated: 2026/09/02 15:42:48 by chguerre         ###   ########.fr       */
+/*   Updated: 2026/09/02 22:30:41 by chguerre         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,22 +26,27 @@ int main(int argc, char **argv, char **envp)
 	int status;
 	int code = 0;
 	char str[] = "Hola, espero que estes bien.\nMis vacaciones van genial.\nLastima tengo que volver al trabajo manana.";
-	
+
 	(void)argc;
 	(void)argv;
-	
+
 	fd = open("temp.txt", O_CREAT | O_WRONLY | O_TRUNC, 0644);
 	if(fd < 0)
 	{
-		perror("Error:-->");
+		perror("Error");
 		exit(1);
 	}
 	write(fd, &str, strlen(str));
-	
+
 	fd2 = open("temp.txt", O_RDONLY);
+	if(fd2 < 0)
+	{
+		perror("Error");
+		exit(1);
+	}
 	unlink("temp.txt");
 	dup2(fd2, STDIN_FILENO);
-	
+
 	close(fd);
 	close(fd2);
 	pid = fork();
